+++ Source/PluginEditor.h
@@ -43,10 +43,14 @@ private:
     juce::ToggleButton playButton;
     juce::ToggleButton overdubButton;
     juce::TextButton clearButton;
+    juce::ToggleButton halfSpeedButton;
+    juce::ToggleButton reverseButton;
     std::unique_ptr<juce::AudioProcessorValueTreeState::ButtonAttachment> recordAttachment;
     std::unique_ptr<juce::AudioProcessorValueTreeState::ButtonAttachment> playAttachment;
     std::unique_ptr<juce::AudioProcessorValueTreeState::ButtonAttachment> overdubAttachment;
     std::unique_ptr<juce::AudioProcessorValueTreeState::ButtonAttachment> clearAttachment;
+    std::unique_ptr<juce::AudioProcessorValueTreeState::ButtonAttachment> halfSpeedAttachment;
+    std::unique_ptr<juce::AudioProcessorValueTreeState::ButtonAttachment> reverseAttachment;
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(SixteenSecondAudioProcessorEditor)
 };